Move Process into PipelineInvoker
