+++ src/NServiceBus.AzureFunctions.Worker.ServiceBus/FunctionEndpoint.cs
@@ -5,9 +5,7 @@
     using System.Threading;
     using System.Threading.Tasks;
     using AzureFunctions.Worker.ServiceBus;
-    using Extensibility;
     using Microsoft.Azure.Functions.Worker;
-    using Transport;
 
     /// <summary>
     /// An NServiceBus endpoint hosted in Azure Function which does not receive messages automatically but only handles
@@ -37,72 +35,8 @@
             await InitializeEndpointIfNecessary(functionContext, CancellationToken.None)
                 .ConfigureAwait(false);
 
-            await Process(body, userProperties, messageId, deliveryCount, replyTo, correlationId, NoTransactionStrategy.Instance, pipeline, cancellationToken)
+            await pipeline.Process(body, userProperties, messageId, deliveryCount, replyTo, correlationId, NoTransactionStrategy.Instance, cancellationToken)
                 .ConfigureAwait(false);
-        }
-
-        internal static async Task Process(
-            byte[] body,
-            IDictionary<string, object> userProperties,
-            string messageId,
-            int deliveryCount,
-            string replyTo,
-            string correlationId,
-            ITransactionStrategy transactionStrategy,
-            PipelineInvoker pipeline,
-            CancellationToken cancellationToken)
-        {
-            body ??= Array.Empty<byte>(); // might be null
-            messageId ??= Guid.NewGuid().ToString("N");
-
-            try
-            {
-                using (var transaction = transactionStrategy.CreateTransaction())
-                {
-                    var transportTransaction = transactionStrategy.CreateTransportTransaction(transaction);
-                    var messageContext = new MessageContext(
-                        messageId,
-                        CreateNServiceBusHeaders(userProperties, replyTo, correlationId),
-                        body,
-                        transportTransaction,
-                        pipeline.ReceiveAddress,
-                        new ContextBag());
-
-                    await pipeline.PushMessage(messageContext, cancellationToken).ConfigureAwait(false);
-
-                    await transactionStrategy.Complete(transaction).ConfigureAwait(false);
-
-                    transaction?.Commit();
-                }
-            }
-            catch (Exception exception)
-            {
-                using (var transaction = transactionStrategy.CreateTransaction())
-                {
-                    var transportTransaction = transactionStrategy.CreateTransportTransaction(transaction);
-                    var errorContext = new ErrorContext(
-                        exception,
-                        CreateNServiceBusHeaders(userProperties, replyTo, correlationId),
-                        messageId,
-                        body,
-                        transportTransaction,
-                        deliveryCount,
-                        pipeline.ReceiveAddress,
-                        new ContextBag());
-
-                    var errorHandleResult = await pipeline.PushFailedMessage(errorContext, cancellationToken).ConfigureAwait(false);
-
-                    if (errorHandleResult == ErrorHandleResult.Handled)
-                    {
-                        await transactionStrategy.Complete(transaction).ConfigureAwait(false);
-
-                        transaction?.Commit();
-                        return;
-                    }
-
-                    throw;
-                }
-            }
         }
 
         async Task InitializeEndpointIfNecessary(FunctionContext functionContext, CancellationToken cancellationToken)
@@ -203,30 +137,6 @@
         /// <inheritdoc />
         public Task Unsubscribe(Type eventType, FunctionContext functionContext, CancellationToken cancellationToken)
             => Unsubscribe(eventType, new UnsubscribeOptions(), functionContext, cancellationToken);
-
-        static Dictionary<string, string> CreateNServiceBusHeaders(IDictionary<string, object> userProperties, string replyTo, string correlationId)
-        {
-            var headers = new Dictionary<string, string>(userProperties.Count);
-
-            foreach (var userProperty in userProperties)
-            {
-                headers[userProperty.Key] = userProperty.Value?.ToString();
-            }
-
-            headers.Remove("NServiceBus.Transport.Encoding");
-
-            if (!string.IsNullOrWhiteSpace(replyTo))
-            {
-                headers.TryAdd(Headers.ReplyToAddress, replyTo);
-            }
-
-            if (!string.IsNullOrWhiteSpace(correlationId))
-            {
-                headers.TryAdd(Headers.CorrelationId, correlationId);
-            }
-
-            return headers;
-        }
 
         readonly Func<FunctionContext, Task<IEndpointInstance>> endpointFactory;
