+++ src/NServiceBus.AzureFunctions.Worker.ServiceBus/Serverless/TransportWrapper/PipelineInvoker.cs
@@ -1,7 +1,10 @@
 ﻿namespace NServiceBus.AzureFunctions.Worker.ServiceBus
 {
+    using System.Collections.Generic;
+    using System;
     using System.Threading;
     using System.Threading.Tasks;
+    using NServiceBus.Extensibility;
     using Transport;
 
     class PipelineInvoker : IMessageReceiver
@@ -10,10 +13,6 @@
         {
             this.baseTransportReceiver = baseTransportReceiver;
         }
-
-        public Task<ErrorHandleResult> PushFailedMessage(ErrorContext errorContext, CancellationToken cancellationToken) => onError(errorContext, cancellationToken);
-
-        public Task PushMessage(MessageContext messageContext, CancellationToken cancellationToken) => onMessage.Invoke(messageContext, cancellationToken);
 
         public Task Initialize(PushRuntimeSettings limitations, OnMessage onMessage, OnError onError,
             CancellationToken cancellationToken)
@@ -24,6 +23,93 @@
                 (_, __) => Task.CompletedTask,
                 (_, __) => Task.FromResult(ErrorHandleResult.Handled),
                 cancellationToken) ?? Task.CompletedTask;
+        }
+
+        public async Task Process(
+            byte[] body,
+            IDictionary<string, object> userProperties,
+            string messageId,
+            int deliveryCount,
+            string replyTo,
+            string correlationId,
+            ITransactionStrategy transactionStrategy,
+            CancellationToken cancellationToken)
+        {
+            body ??= Array.Empty<byte>(); // might be null
+            messageId ??= Guid.NewGuid().ToString("N");
+
+            try
+            {
+                using (var transaction = transactionStrategy.CreateTransaction())
+                {
+                    var transportTransaction = transactionStrategy.CreateTransportTransaction(transaction);
+                    var messageContext = new MessageContext(
+                        messageId,
+                        CreateNServiceBusHeaders(userProperties, replyTo, correlationId),
+                        body,
+                        transportTransaction,
+                        ReceiveAddress,
+                        new ContextBag());
+
+                    await onMessage(messageContext, cancellationToken).ConfigureAwait(false);
+
+                    await transactionStrategy.Complete(transaction).ConfigureAwait(false);
+
+                    transaction?.Commit();
+                }
+            }
+            catch (Exception exception)
+            {
+                using (var transaction = transactionStrategy.CreateTransaction())
+                {
+                    var transportTransaction = transactionStrategy.CreateTransportTransaction(transaction);
+                    var errorContext = new ErrorContext(
+                        exception,
+                        CreateNServiceBusHeaders(userProperties, replyTo, correlationId),
+                        messageId,
+                        body,
+                        transportTransaction,
+                        deliveryCount,
+                        ReceiveAddress,
+                        new ContextBag());
+
+                    var errorHandleResult = await onError.Invoke(errorContext, cancellationToken).ConfigureAwait(false);
+
+                    if (errorHandleResult == ErrorHandleResult.Handled)
+                    {
+                        await transactionStrategy.Complete(transaction).ConfigureAwait(false);
+
+                        transaction?.Commit();
+                        return;
+                    }
+
+                    throw;
+                }
+            }
+
+            static Dictionary<string, string> CreateNServiceBusHeaders(IDictionary<string, object> userProperties, string replyTo, string correlationId)
+            {
+                var headers = new Dictionary<string, string>(userProperties.Count);
+
+                foreach (var userProperty in userProperties)
+                {
+                    headers[userProperty.Key] = userProperty.Value?.ToString();
+                }
+
+                headers.Remove("NServiceBus.Transport.Encoding");
+
+                if (!string.IsNullOrWhiteSpace(replyTo))
+                {
+                    headers.TryAdd(Headers.ReplyToAddress, replyTo);
+                }
+
+                if (!string.IsNullOrWhiteSpace(correlationId))
+                {
+                    headers.TryAdd(Headers.CorrelationId, correlationId);
+                }
+
+                return headers;
+            }
         }
 
         public Task StartReceive(CancellationToken cancellationToken) => Task.CompletedTask;
